+++ AWSAppSyncClient/Internal/AWSPerformMutationQueue.swift
@@ -17,11 +17,9 @@ import Foundation
 
 final class AWSPerformMutationQueue {
 
-    private unowned let appSyncClient: AWSAppSyncClient
-    private let networkClient: AWSNetworkTransport
-    private let snapshotProcessController: SnapshotProcessController
-
-    private var persistentCache: AWSMutationCache?
+    private weak var appSyncClient: AWSAppSyncClient?
+    private weak var networkClient: AWSNetworkTransport?
+    private let persistentCache: AWSMutationCache?
 
     private let operationQueue: OperationQueue
     private let handlerQueue: DispatchQueue
@@ -30,94 +28,41 @@ final class AWSPerformMutationQueue {
         appSyncClient: AWSAppSyncClient,
         networkClient: AWSNetworkTransport,
         handlerQueue: DispatchQueue = .main,
-        snapshotProcessController: SnapshotProcessController,
-        fileURL: URL? = nil) {
+        reachabiltyChangeNotifier: NetworkReachabilityNotifier?,
+        cacheFileURL: URL? = nil) {
+
         self.appSyncClient = appSyncClient
         self.networkClient = networkClient
-        self.snapshotProcessController = snapshotProcessController
-
         self.handlerQueue = handlerQueue
 
         self.operationQueue = OperationQueue()
         self.operationQueue.name = "com.amazonaws.service.appsync.MutationQueue"
         self.operationQueue.maxConcurrentOperationCount = 1
 
-        if let fileURL = fileURL {
+        if let cacheFileURL = cacheFileURL {
             do {
-                persistentCache = try AWSMutationCache(fileURL: fileURL)
+                persistentCache = try AWSMutationCache(fileURL: cacheFileURL)
 
                 operationQueue.addOperation { [weak self] in
-                    self?.loadMutations()
-                }
-            } catch {
-                debugPrint("persistentCache initialization error: \(error)")
-            }
-        }
-
-        self.suspendOrResumeQueue()
-    }
-
-    // MARK: Offline Mutations
-
-    private func loadMutations() {
-        do {
-            guard let mutations = try persistentCache?.getStoredMutationRecordsInQueue() else {
-                return
-            }
-
-            for mutation in mutations {
-                let operation = AWSPerformOfflineMutationOperation(
-                    appSyncClient: appSyncClient,
-                    networkClient: networkClient,
-                    handlerQueue: handlerQueue,
-                    mutation: mutation)
-
-                operation.operationCompletionBlock = { [weak self] operation, error in
-                    let identifier = operation.mutation.recordIdentitifer
-
                     do {
-                        try self?.deleteOfflineMutation(withIdentifier: identifier)
+                        try self?.loadMutations()
                     } catch {
-                        debugPrint("deleteOfflineMutation error: \(error)")
+                        print("Error loading mutations: \(error)")
                     }
                 }
-
-                operationQueue.addOperation(operation)
+            } catch {
+                persistentCache = nil
+                debugPrint("persistentCache initialization error: \(error)")
             }
-        } catch {
-            debugPrint("\(#function) error: \(error)")
-        }
-    }
-
-    private func save<Mutation: GraphQLMutation>(_ mutation: Mutation) throws -> AWSAppSyncMutationRecord? {
-        guard let persistentCache = persistentCache else { return nil }
-
-        let requestBody = AWSRequestBuilder.requestBody(from: mutation)
-        let data = try JSONSerializationFormat.serialize(value: requestBody)
-
-        let offlineMutation = AWSAppSyncMutationRecord()
-
-        if let s3Object = AWSRequestBuilder.s3Object(from: mutation.variables) {
-            offlineMutation.type = .graphQLMutationWithS3Object
-            offlineMutation.s3ObjectInput = s3Object
+        } else {
+            persistentCache = nil
         }
 
-        offlineMutation.data = data
-        offlineMutation.contentMap = mutation.variables
-        offlineMutation.jsonRecord = mutation.variables?.jsonObject
-        offlineMutation.recordState = .inQueue
-        offlineMutation.operationString = Mutation.operationString
-
-        try persistentCache.saveMutationRecord(record: offlineMutation)
-
-        return offlineMutation
+        self.suspendOrResumeQueue(reachabiltyChangeNotifier: reachabiltyChangeNotifier)
+        reachabiltyChangeNotifier?.add(watcher: self)
     }
 
-    private func deleteOfflineMutation(withIdentifier identifier: String) throws {
-        try persistentCache?.deleteMutationRecord(withIdentifier: identifier)
-    }
-
-    //
+    // MARK: - Queue operations
 
     func add<Mutation: GraphQLMutation>(
         _ mutation: Mutation,
@@ -164,20 +109,83 @@ final class AWSPerformMutationQueue {
         operationQueue.isSuspended = false
     }
 
-    private func suspendOrResumeQueue() {
-        if snapshotProcessController.isNetworkReachable {
+    private func suspendOrResumeQueue(reachabiltyChangeNotifier: NetworkReachabilityNotifier?) {
+        if reachabiltyChangeNotifier?.isNetworkReachable ?? false {
             resume()
         } else {
             suspend()
         }
     }
+
+    // MARK: Offline Mutations
+
+    private func loadMutations() throws {
+        do {
+            guard let mutations = try persistentCache?.getStoredMutationRecordsInQueue() else {
+                return
+            }
+
+            for mutation in mutations {
+                let operation = AWSPerformOfflineMutationOperation(
+                    appSyncClient: appSyncClient,
+                    networkClient: networkClient,
+                    handlerQueue: handlerQueue,
+                    mutation: mutation)
+
+                operation.operationCompletionBlock = { [weak self] operation, error in
+                    let identifier = operation.mutation.recordIdentitifer
+
+                    do {
+                        try self?.deleteOfflineMutation(withIdentifier: identifier)
+                    } catch {
+                        debugPrint("deleteOfflineMutation error: \(error)")
+                    }
+                }
+
+                operationQueue.addOperation(operation)
+            }
+        } catch {
+            debugPrint("\(#function) error: \(error)")
+        }
+    }
+
+    private func save<Mutation: GraphQLMutation>(_ mutation: Mutation) throws -> AWSAppSyncMutationRecord? {
+        guard let persistentCache = persistentCache else {
+            return nil
+        }
+
+        let requestBody = AWSRequestBuilder.requestBody(from: mutation)
+        let data = try JSONSerializationFormat.serialize(value: requestBody)
+
+        let offlineMutation = AWSAppSyncMutationRecord()
+
+        if let s3Object = AWSRequestBuilder.s3Object(from: mutation.variables) {
+            offlineMutation.type = .graphQLMutationWithS3Object
+            offlineMutation.s3ObjectInput = s3Object
+        }
+
+        offlineMutation.data = data
+        offlineMutation.contentMap = mutation.variables
+        offlineMutation.jsonRecord = mutation.variables?.jsonObject
+        offlineMutation.recordState = .inQueue
+        offlineMutation.operationString = Mutation.operationString
+
+        try persistentCache.saveMutationRecord(record: offlineMutation)
+
+        return offlineMutation
+    }
+
+    private func deleteOfflineMutation(withIdentifier identifier: String) throws {
+        try persistentCache?.deleteMutationRecord(withIdentifier: identifier)
+    }
+
 }
 
 // MARK: - NetworkConnectionNotification
 
-extension AWSPerformMutationQueue: NetworkConnectionNotification {
+extension AWSPerformMutationQueue: NetworkReachabilityWatcher {
 
-    func onNetworkAvailabilityStatusChanged(isEndpointReachable: Bool) {
+    func onNetworkReachabilityChanged(isEndpointReachable: Bool) {
         if isEndpointReachable {
             resume()
         } else {